+++ src/views/apps/cves/targetDetectionListCVEAll.vue
@@ -34,9 +34,9 @@
                             <span class="me-2">{{ selectedIds.length }}</span>Selected
                         </div>
                         <el-tooltip class="box-item" effect="dark" hide-after="0" content="Xóa" placement="top">
-                            <button type="button" data-bs-target="#kt_modal_delete" data-bs-toggle="modal"
-                                class="btn btn-danger btn-sm " :disabled="disabled">
-                                Delete Selected
+                            <button type="button" @click="deleteSelectd()" class="btn btn-danger btn-sm "
+                                :disabled="disabled">
+                                Xóa mục đã chọn
                             </button>
                         </el-tooltip>
                     </div>
@@ -218,35 +218,7 @@
         </div>
     </div>
 
-    <div class="modal fade" tabindex="-1" id="kt_modal_delete" ref="ModalDelete" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered">
-            <div class="modal-content">
 
-                <div class="modal-header">
-                    <h3 class="modal-title">Xác nhận xóa scan</h3>
-
-                    <div class="btn btn-icon btn-sm btn-active-light-primary ms-2" data-bs-dismiss="modal"
-                        aria-label="Close">
-                        <span class="svg-icon svg-icon-2x"></span>
-                    </div>
-                </div>
-                <div class="modal-body">
-                    <p style="font-size: 16px;">Bạn có chắc chắn muốn xóa <span v-if="selectedIds.length > 0"
-                            class="fw-bold" style="color:red;">{{ selectedIds.length
-                            }}</span> bản ghi này không?
-                    </p>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-sm  btn-light" data-bs-dismiss="modal">
-                        Hủy bỏ
-                    </button>
-                    <button type="button" class="btn btn-sm  btn-primary" @click.passive="deleteFewSubscriptions()">
-                        Đồng ý
-                    </button>
-                </div>
-            </div>
-        </div>
-    </div>
 </template>
 
 <style lang="scss">
@@ -270,11 +242,13 @@ import ApiService from "@/core/services/ApiService";
 import { hideModal } from "@/core/helpers/dom";
 import { ErrorMessage, Field, Form as VForm } from "vee-validate";
 import { vue3Debounce } from 'vue-debounce';
-import { ElMessage } from 'element-plus'
 import Fillter from "@/views/apps/targets/filterTargetDectionListCVE.vue";
 import axios from 'axios'
 import * as Yup from "yup";
 import Swal from "sweetalert2/dist/sweetalert2.js";
+import { ElMessage, ElMessageBox } from 'element-plus'
+import { Delete } from '@element-plus/icons-vue'
+import { markRaw } from 'vue'
 
 import { Modal } from "bootstrap";
 import dayjs from 'dayjs';
@@ -478,8 +452,8 @@ export default defineComponent({
                 });
         }
 
-                // tính thời gian
-                const eventTime = ref<number | any>('30000');
+        // tính thời gian
+        const eventTime = ref<number | any>('30000');
         let intervalId: any;
         const startTimer = () => {
             intervalId = setInterval(() => {
@@ -500,12 +474,27 @@ export default defineComponent({
         });
 
         const selectedIds = ref<Array<number>>([]);
-        const deleteFewSubscriptions = () => {
-            deleteSubscription(selectedIds.value);
+        const deleteSelectd = () => {
+            ElMessageBox.confirm(
+                'Tập tin sẽ được xóa vĩnh viễn. Tiếp tục?',
+                'Xác nhận xóa',
+                {
+                    confirmButtonText: 'Đồng ý',
+                    cancelButtonText: 'Hủy bỏ',
+                    type: 'warning',
+                    icon: markRaw(Delete)
+                }
+            )
+                .then(() => {
+                    deleteSubscription(selectedIds.value);
+                })
+                .catch(() => {
+
+                })
         };
         const disabled = ref<boolean>(false);
 
-        const ModalDelete = ref<null | HTMLElement>(null);
+        
         const deleteSubscription = (ids: Array<number>) => {
             let formData = {
                 'id': ids
@@ -574,7 +563,7 @@ export default defineComponent({
                 },
             }).then(() => {
                 hideModal(newTargetGroupModalRef.value);
-                hideModal(ModalDelete.value);
+                
                 hideModal(ModalDetail.value);
             });
         }
@@ -642,8 +631,8 @@ export default defineComponent({
             sort,
             onItemSelect,
             selectedIds,
-            deleteFewSubscriptions,
-            deleteSubscription,
+            deleteSelectd,
+            
             getAssetPath,
 
             // validate
@@ -657,7 +646,7 @@ export default defineComponent({
             newTargetGroupModalRef,
             // handleClick,
             errors,
-            ModalDelete,
+            
             discardButtonRef,
             // detials
             ModalDetail,
